pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fanuc-ml"
version = "0.1.0"
description = "Machine Learning Framework for FANUC Robot Arm Positioning"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "FANUC ML Team", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
    "torch>=1.10.0",
    "numpy>=1.20.0",
    "pybullet>=3.2.1",
    "gymnasium>=0.28.1",
    "stable-baselines3>=2.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "imageio>=2.22.0",
    "pillow>=9.0.0",
]

[project.optional-dependencies]
directml = [
    "torch-directml>=0.2.0.dev230426"
]

[project.urls]
"Homepage" = "https://github.com/yourusername/fanuc-ml"
"Bug Tracker" = "https://github.com/yourusername/fanuc-ml/issues"

[project.scripts]
fanuc-train = "main:main"
fanuc-eval = "tools.run_eval:main"  
fanuc-test = "tools.test_install:main"

[tool.setuptools]
packages = ["src"]
package-dir = {"" = "src", "fanuc_tools" = "tools"}
py-modules = ["main"] 